首页

+++ src/main/java/com/coldraincn/laimihui/entity/ProductDetail.java
@@ -10,9 +10,10 @@ import java.util.List;
 public class ProductDetail implements Serializable {
     private static final long serialVersionUID = 1L;
 
+
     /**
      * status : OK
-     * data : {"coverId":"http://api.17yun.com.cn/upload/20170731/533821041184991.jpg","imglist":["http://dev.wecity.co/upload/20170731/533823460732199.jpg","http://dev.wecity.co/upload/20170904/245930890258055.jpg","http://dev.wecity.co/upload/20170904/245933679275300.jpg"],"marketPrice":199,"objectId":533840800783484,"postPrice":0,"productInfo":"<p><img src=\"/task/ueditor/readImage.do?path=/home/thinker/wwwroot/dev.wecity.co/upload/image/20170904/1504509989015007053.jpg\" title=\"1504509989015007053.jpg\" style=\"width:100%;\" alt=\"微信图片_20170831101644.jpg\"/><\/p>","productKey":"新鲜果蔬","productName":"测试收货地址","productQuota":-1,"productStatus":"3","productStatusName":"进行中","servicePhoneNo":"4000088910","shareDescribe":"优惠券测试3","shareImg":"http://dev.wecity.co/upload/20170731/533834843070655.jpg","shareTitle":"优惠券测试3","showCostPrice":199,"showSalePrice":0.01,"singleSellPrice":0.02,"taobaoPrice":199}
+     * data : {"coverId":"http://api.17yun.com.cn/upload/20170704/19080749993030166.jpg","imglist":["http://dev.wecity.co/upload/20170904/245788401617884.jpg","http://dev.wecity.co/upload/20170904/245792444547730.jpg","http://dev.wecity.co/upload/20170904/245797451505722.jpg"],"marketPrice":0.1,"objectId":537580830664190,"postPrice":0,"productInfo":"<p><span style=\"font-family: 华文细黑; color: rgb(0, 0, 0); font-size: 20px;\">此项非常规商品，用于较偏远地区邮费差价。拍前请联系客服，请勿自行购买。<\/span><\/p><p><span style=\"font-family: 华文细黑; color: rgb(0, 0, 0); font-size: 20px;\"><img src=\"http://dev.wecity.co/task/ueditor/readImage.do?path=/home/thinker/wwwroot/dev.wecity.co/upload/image/20170904/1504509793370068373.jpg\" title=\"1504509793370068373.jpg\" style=\"width:100%;\" alt=\"微信图片_20170831103044.jpg\"/><\/span><\/p>","productKey":"其他，此项非常规商品，拍前请联系客服，请勿自行购买。邮费，其他","productName":"补拍邮费","productQuota":-1,"productStatus":"3","productStatusName":"进行中","servicePhoneNo":"4000082910","shareDescribe":" 此项非常规商品，请联系客服后购买","shareImg":"http://dev.wecity.co/upload/20170710/19579510208129592.jpg","shareTitle":"邮费补拍","showCostPrice":0.1,"showSalePrice":0.01,"singleSellPrice":0.1,"taobaoPrice":0.1}
      */
 
     private String status;
@@ -34,34 +35,34 @@ public class ProductDetail implements Serializable {
         this.data = data;
     }
 
-    public static class DataBean  implements Serializable {
+    public static class DataBean implements Serializable{
         private static final long serialVersionUID = 1L;
         /**
-         * coverId : http://api.17yun.com.cn/upload/20170731/533821041184991.jpg
-         * imglist : ["http://dev.wecity.co/upload/20170731/533823460732199.jpg","http://dev.wecity.co/upload/20170904/245930890258055.jpg","http://dev.wecity.co/upload/20170904/245933679275300.jpg"]
-         * marketPrice : 199
-         * objectId : 533840800783484
+         * coverId : http://api.17yun.com.cn/upload/20170704/19080749993030166.jpg
+         * imglist : ["http://dev.wecity.co/upload/20170904/245788401617884.jpg","http://dev.wecity.co/upload/20170904/245792444547730.jpg","http://dev.wecity.co/upload/20170904/245797451505722.jpg"]
+         * marketPrice : 0.1
+         * objectId : 537580830664190
          * postPrice : 0
-         * productInfo : <p><img src="/task/ueditor/readImage.do?path=/home/thinker/wwwroot/dev.wecity.co/upload/image/20170904/1504509989015007053.jpg" title="1504509989015007053.jpg" style="width:100%;" alt="微信图片_20170831101644.jpg"/></p>
-         * productKey : 新鲜果蔬
-         * productName : 测试收货地址
+         * productInfo : <p><span style="font-family: 华文细黑; color: rgb(0, 0, 0); font-size: 20px;">此项非常规商品，用于较偏远地区邮费差价。拍前请联系客服，请勿自行购买。</span></p><p><span style="font-family: 华文细黑; color: rgb(0, 0, 0); font-size: 20px;"><img src="http://dev.wecity.co/task/ueditor/readImage.do?path=/home/thinker/wwwroot/dev.wecity.co/upload/image/20170904/1504509793370068373.jpg" title="1504509793370068373.jpg" style="width:100%;" alt="微信图片_20170831103044.jpg"/></span></p>
+         * productKey : 其他，此项非常规商品，拍前请联系客服，请勿自行购买。邮费，其他
+         * productName : 补拍邮费
          * productQuota : -1
          * productStatus : 3
          * productStatusName : 进行中
-         * servicePhoneNo : 4000088910
-         * shareDescribe : 优惠券测试3
-         * shareImg : http://dev.wecity.co/upload/20170731/533834843070655.jpg
-         * shareTitle : 优惠券测试3
-         * showCostPrice : 199
+         * servicePhoneNo : 4000082910
+         * shareDescribe :  此项非常规商品，请联系客服后购买
+         * shareImg : http://dev.wecity.co/upload/20170710/19579510208129592.jpg
+         * shareTitle : 邮费补拍
+         * showCostPrice : 0.1
          * showSalePrice : 0.01
-         * singleSellPrice : 0.02
-         * taobaoPrice : 199
+         * singleSellPrice : 0.1
+         * taobaoPrice : 0.1
          */
 
         private String coverId;
-        private int marketPrice;
+        private double marketPrice;
         private long objectId;
-        private int postPrice;
+        private double postPrice;
         private String productInfo;
         private String productKey;
         private String productName;
@@ -72,10 +73,10 @@ public class ProductDetail implements Serializable {
         private String shareDescribe;
         private String shareImg;
         private String shareTitle;
-        private int showCostPrice;
+        private double showCostPrice;
         private double showSalePrice;
         private double singleSellPrice;
-        private int taobaoPrice;
+        private double taobaoPrice;
         private List<String> imglist;
 
         public String getCoverId() {
@@ -86,11 +87,11 @@ public class ProductDetail implements Serializable {
             this.coverId = coverId;
         }
 
-        public int getMarketPrice() {
+        public double getMarketPrice() {
             return marketPrice;
         }
 
-        public void setMarketPrice(int marketPrice) {
+        public void setMarketPrice(double marketPrice) {
             this.marketPrice = marketPrice;
         }
 
@@ -102,11 +103,11 @@ public class ProductDetail implements Serializable {
             this.objectId = objectId;
         }
 
-        public int getPostPrice() {
+        public double getPostPrice() {
             return postPrice;
         }
 
-        public void setPostPrice(int postPrice) {
+        public void setPostPrice(double postPrice) {
             this.postPrice = postPrice;
         }
 
@@ -190,11 +191,11 @@ public class ProductDetail implements Serializable {
             this.shareTitle = shareTitle;
         }
 
-        public int getShowCostPrice() {
+        public double getShowCostPrice() {
             return showCostPrice;
         }
 
-        public void setShowCostPrice(int showCostPrice) {
+        public void setShowCostPrice(double showCostPrice) {
             this.showCostPrice = showCostPrice;
         }
 
@@ -214,11 +215,11 @@ public class ProductDetail implements Serializable {
             this.singleSellPrice = singleSellPrice;
         }
 
-        public int getTaobaoPrice() {
+        public double getTaobaoPrice() {
             return taobaoPrice;
         }
 
-        public void setTaobaoPrice(int taobaoPrice) {
+        public void setTaobaoPrice(double taobaoPrice) {
             this.taobaoPrice = taobaoPrice;
         }
